feat: 불필요한 조건문 및 로직 제거
